Revert "Organized code and renamed classes"
