--- AUTHORS
@@ -1,3 +1,0 @@
-Christophe Servan <c.servan@qwantresearch.com>
-Noel Martin <n.martin@qwantresearch.com>
-Pierre Jackman <p.jackman@qwantresearch.com>
