+++ README.md
@@ -1,24 +1,33 @@
-Katanoisi
-=========
+Qclassification
+===============
 
-A new C++ API for Intention classification at Qwant Research.
-The API is based on `fasttext`.
+A new CPP API for Intention classification at Qwant Research.
+The API is based on fasttext.
 
 ## Installation
+
+### Pre-requieres
+* install fasttext (https://github.com/facebookresearch/fastText.git)
+* install forked version of pistache (https://github.com/QwantResearch/pistache.git)
+* install nlohmann json (https://github.com/nlohmann/json.git)
+* install qnlp-toolkit (https://github.com/QwantResearch/qnlp-toolkit.git)
+
+
+### Installation
+
+
 ```
-git clone --recursive https://github.com/QwantResearch/Katanoisi.git 
-docker build .
+  git clone https://github.com/QwantResearch/qclassification.git 
+  git checkout cpp
+  mkdir -pv build && cd build && cmake .. && make 
 ``` 
 
-## Launch the API
-```
-./katanoisi --model-config <filename> [--port <port>] [--threads <nthreads>] [--debug]
 
---model-config (-f)      config file in which all models are described (REQUIRED)
---port (-p)              port to use (default 9009)
---threads (-t)           number of threads (default 1)
---debug (-d)             debug mode (default false)
---help (-h)              Show this message
+## Launch the API
+
+Launch the API:
+```
+  ./qclass_server [#port] [#threads] [config filename]
 ```
 
 ## Licencing
@@ -28,3 +37,4 @@ Copyright 2019 Qwant Research. Licensed under the terms of the Apache 2.0 licens
 Contact:
  - christophe[dot]servan[at]qwantresearch[dot]com
  - n[dot]martin[at]qwantresearch[dot]com
+
